Huecost not calculating is the issue

+++ AStar.h
@@ -7,7 +7,7 @@
 //represents nodes of a graph
 struct Node{
     int x, y; //coordinates
-    int f, g, h; //A* algorithm values
+    double f, g, h; //A* algorithm values
     /*
     g is cost to reach a node from the start node
     h is the estimated cost to reach the goal node from the current node
@@ -21,7 +21,7 @@ struct Node{
 };
 
 //Heuristic Implementation
-int HeuCost(int sx, int sy, int gx, int gy);
+double HeuCost(int sx, int sy, int gx, int gy);
 
 //A* Algorithm Call
 std::vector<Node> FindPath(const std::vector<std::vector<int>>& graph, const Node& start, const Node& goal);
